Finished the implementation of interrupting inputs and programs. It's now working on the basement wing zones. Removed the race condition that was happening earlier. It was because the Start method was setting the running indicator asynchronously and the Stop method was checking it synchronously. This was causing the running indicator to always be the wrong value.
Also changed all tests to NUnit because XUnit is causing many issues including not being able to run tests individually. It's also causing other weirdnesses with VS 2015 and R# 9.0 I think. Need to fix the ForceStop and CooperativeStop tests because they fail when run with the other tests. They pass when running individually, so it sounds like badly-written tests.

Added a new program called BlinkColor that blinks a certain color.