Lots of refactor. Fixed Shimmer bug where it didn't power down completely. It was being caused by an incorrect implementation of Force Stop. Force Stop is now implemented outside out ZoneProgram (it was previously inside ZoneProgram). The problem was that the force stop call was not being tracked by Fade and other methods that need to do live tracking of the force stop flag. The way it's implemented now, force stop does not kill a thread the way it was originally intended, but rather request the internal methods to stop immediately, which are now to be designed to follow this paradigm of being able to stop immediately.  Fade has already been modified to act this way. When a regular stop is called on Shimmer, it waits until the current lights finish fading to black. When a force stop is called, it stops immediately. Because the Force Stop Flag is only used at the program's subclass level, it was removed from ZoneProgram. ZoneProgram itself doesn't do anything with the ForceStoppable flag, it delegates that responsibility to the subclass. So why have it in ZP at all? This also eliminated the weird inheritance that I was trying to squeeze out of ForceStoppable.
Clock: This is a big one. In trying to come up with a good clocking mechanism for ZL, I've had to experiment with a lot of different things. This includes a library called MicroTimer and MicroStopwatch (in the eponymous file). Turns out these libraries were too processor exhaustive. I also built a clock myself using a Timer and Stopwatch (in eponymous files). These were not accurate enough. Then I tried modifying the MicroClock to try to achieve close to 1 ms accuracy. Currently the MilliTimer can achieve accuracy of about 10ms +/- 2ms (these number might vary from computer to computer and trial to trial, but this is where my results were averaging). So I will proceed with MilliTimer as the clocking mechanism for Raindrops for now.

Extracted some of the functionality relating to graphics, drawing, and animation into their own namespaces/classes. Added a SetInput method that searches both Zones and ProgramSets to execute the correct method behind the scenes.